e5-1650: auto generate

+++ internal/testfixture/ccgo_linux_386.go
@@ -15714,7 +15714,7 @@ func _test_getrusage(tls *libc.TLS, clientData uintptr, interp uintptr, objc int
 	var _ /* buf at bp+0 */ [1024]int8
 	var _ /* r at bp+1024 */ Trusage
 	libc.Xmemset(tls, bp+1024, 0, uint32(72))
-	libc.Xgetrusage(tls, int32(_RUSAGE_SELF), bp+1024)
+	libc.Xgetrusage(tls, _RUSAGE_SELF, bp+1024)
 	x_sqlite3_snprintf(tls, int32(1024), bp, __ccgo_ts+4091, libc.VaList(bp+1104, (*(*Trusage)(unsafe.Pointer(bp + 1024))).Fru_utime.Ftv_sec, (*(*Trusage)(unsafe.Pointer(bp + 1024))).Fru_utime.Ftv_usec, (*(*Trusage)(unsafe.Pointer(bp + 1024))).Fru_stime.Ftv_sec, (*(*Trusage)(unsafe.Pointer(bp + 1024))).Fru_stime.Ftv_usec, (*(*Trusage)(unsafe.Pointer(bp + 1024))).F__ccgo6_32.Fru_minflt, (*(*Trusage)(unsafe.Pointer(bp + 1024))).F__ccgo7_36.Fru_majflt))
 	libtcl8_6.XTcl_SetObjResult(tls, interp, libtcl8_6.XTcl_NewStringObj(tls, bp, -int32(1)))
 	return m_TCL_OK
@@ -45770,7 +45770,7 @@ func _postToParent(tls *libc.TLS, p uintptr, pScript uintptr) {
 	(*TEvalEvent)(unsafe.Pointer(pEvent)).FzScript = pEvent + 1*16
 	libc.Xmemcpy(tls, (*TEvalEvent)(unsafe.Pointer(pEvent)).FzScript, zMsg, uint32(*(*int32)(unsafe.Pointer(bp))+int32(1)))
 	(*TEvalEvent)(unsafe.Pointer(pEvent)).Finterp = (*TSqlThread)(unsafe.Pointer(p)).Finterp
-	libtcl8_6.XTcl_ThreadQueueEvent(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent, pEvent, int32(_TCL_QUEUE_TAIL))
+	libtcl8_6.XTcl_ThreadQueueEvent(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent, pEvent, _TCL_QUEUE_TAIL)
 	libtcl8_6.XTcl_ThreadAlert(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent)
 }
 
@@ -45914,7 +45914,7 @@ func _sqlthread_parent(tls *libc.TLS, clientData TClientData, interp uintptr, ob
 	(*TEvalEvent)(unsafe.Pointer(pEvent)).FzScript = pEvent + 1*16
 	libc.Xmemcpy(tls, (*TEvalEvent)(unsafe.Pointer(pEvent)).FzScript, zMsg, uint32(*(*int32)(unsafe.Pointer(bp))+int32(1)))
 	(*TEvalEvent)(unsafe.Pointer(pEvent)).Finterp = (*TSqlThread)(unsafe.Pointer(p)).Finterp
-	libtcl8_6.XTcl_ThreadQueueEvent(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent, pEvent, int32(_TCL_QUEUE_TAIL))
+	libtcl8_6.XTcl_ThreadQueueEvent(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent, pEvent, _TCL_QUEUE_TAIL)
 	libtcl8_6.XTcl_ThreadAlert(tls, (*TSqlThread)(unsafe.Pointer(p)).Fparent)
 	return m_TCL_OK
 }
